Reused useEmitterEvents
